Add take/3
This could be implemented in terms of length/2 and prefix/2 but having
it together in a single predicate is useful.  I considered naming it
something like prefix_n/3 since that describes the relationship
better, but settled on take/3 since it's short and mirrors the name
used in many functional languages.