Add drop/3
Because drop/3 allows one to drop more elements than are present in a
list, drop/3 has one slightly unusual mode.  If the count is greater
than 0 and the third argument is an empty list, the predicate can
generate multiple solutions.

SWI-Prolog's notion of types is flexible enough that I could encode
this mode in the structured documentation.  I also included it in the
example to clarify exactly when this happens.