Add xfy_list/3
This predicate was originally an implementation detail inside the
function_expansion module.  It has proven useful outside that context
so I want it available in a pack.

I debated whether I should create a new pack called term_util to house
this predicate.  Since the predicate deals with lists and terms, it
seemed relevant in either place.  list_util already exists so I
settled on that.