Implement minimum and maximum predicates
It's often convenient to find the smallest or largest term in a list.
Without dedicated predicates, the easiest way is to sort the list and
choose the first (or last) element.  This requires more computation
than a simple list traversal.

These predicates are also convenient when working with library(func)
since they can be used as functions.