Implement lines/2
It's often convenient to work with a file as if it were just a list of
lines.  However, we don't want to load the entire file into memory.

Unfortunately, this uses the same stream repositioning trick used by
library(pure_input) so it only works on certain kinds of streams.