Remove split_at/4 dangling choice point
When the first argument of split_at/4 was larger than the length of the
second argument, there was a dangling choice point.  Backtracking into
this choice point produced no new solutions.

For example,

  ?- split_at(1000, [a,b,c], [a,b,c], []).
  true ;
  false.

By using an auxiliary predicate (split_at_/4) with the first two
arguments swapped, we can leverage first argument indexing to eliminate
the dangling choice point.  This avoids the time and memory cost of
creating the choice point.  It also avoids unnecessary backtracking.

You can see the improved API experience in the test cases: we no longer
need to wrap with aggregate_all/3.