Implement take/3 in terms of split_at/4
These two predicates have very similar code.  In this case, split_at/4
performs no extra work compared to take/3.  That's because the 4th
argument (whose value we ignore) is just a list's tail which is
essentially free to calculate.

I did not implement drop/3 in terms of split_at/4.  In that case, the
ignored 4th argument is a brand new list, constructed from scratch.
Creating a list that we'll immediately discard requires work and creates
extra garbage.