Call our internal iterator/stream thing "flow"
The word "stream" is perhaps the traditional name for something like, but
it's already taken by Prolog's IO streams.  Using the same word is
confusing.

The word "iterator" is common in object oriented languages for something
that has a next() method.  I'm not sure if that's exactly what we have
here.  It also comes with lots of baggage from those languages.

The word "flow" seems nice since it portrays a visual metaphor similar
to a stream.  In neither case, can one typically access the past
elements.  They "flow" past and are gone.