Modifying lazy_maplist/3 to be lazy in both directions.
I've added an extra freeze mandating that both Xs and Ys be bound
before proceeding to evaluate. I'm aware that this is a little awkward
at first, but I believe that this ultimately allows one more control
over how lazy they want maplist/3 to be.

That being said, I'm completely open to a revision if there's a better
way, or a reversion if this is perhaps deemed an unnecessary feature.