flow: allow next/2 to fail
In some circumstances a flow cannot determine, a priori, whether it has
produced all elements.  It must try to fetch the next value and then
realize it's not possible.  To account for this case, allow next/2 to
fail and mark the nblist as end_of_file at that moment.

This code path is currently not used since fetching lines from a stream
can predict future values with at_end_of_stream.