Minor adjustment to documentation
group_by/3 works in two different modes.  Each mode has different
determinism.  When given a list, we can always create some groups.
However, when given some groups we can only create a list of the
original groups meet the equality constraint.  The current
implementation doesn't make this distinction but for the predicate to be
truly logical we should make that distinction in the documentation.

See #7