Make sure tests trigger the right mode
In the tests for group_by/3 make sure that the unknown argument starts
off as a variable.  This makes certain that providing information for
the output argument isn't biasing the results towards correctness.

This was already done for the backwards mode.  This commit does it for
the forwards mode too.

See #7