Tail recursive group_by/3
The previous implementation of group_by/3 was not tail recursive so it
consumed a lot of memory when grouping really large lists.  This new
implementation is tail recursive so it keeps memory consumption constant
regardless of list size.

As an added bonus, this version runs backwards and forwards without an
explicit call to flatten/2.  It also causes the test case
backwards_but_wrong_equality to pass.

See #7