span/5: change semantics for an empty result
span/5 describes its results using a difference list.  A difference list
is a pair of variables for which the "difference" between them describes
the list we care about.  For span/5 that pair of variables is named
Prefix and Tail.  An empty difference list is communicated by Prefix and
Tail being unified.  Since they're the same, the difference between them
is 0 (ie, empty).

This commit changes the tests and code to implement this semantics.  The
"chained" test indicates a common use case where this semantics is
necessary.

See #22