split/3: support delimiter in separated elements
When calling split/3 in mode split(-,+,+), we now allow the third
argument to contain the delimiter within it.  This used to be
prohibited since there's no way to invoke split(+,+,-) to produce that
output.  However, this new behavior is useful, unambiguous and cleaner
to implement.

I removed the "optimization" comment because, without a call to dif/2,
the cut is no longer an optimization.  It's actually necessary to
obtain correct results.  The cut still improves performance, but
that's no longer its only purpose.

Close #30